Fix dbt model errors and test failures in mart_hygiene_retention
- Fix round() function type errors in mart_hygiene_retention.sql and mart_appointment_summary.sql
  * mart_hygiene_retention.sql (line 155): Cast entire expression to numeric before round()
    - Changed: round(ha.avg_hygiene_production::numeric / nullif(ha.avg_hygiene_duration, 0) * 60, 2)
    - To: round((ha.avg_hygiene_production::numeric / nullif(ha.avg_hygiene_duration, 0) * 60)::numeric, 2)
  * mart_appointment_summary.sql (line 265): Cast denominator to numeric before division
    - Changed: round((ae.completed_production::numeric / nullif(ae.completed_appointments, 0))::numeric, 2)
    - To: round((ae.completed_production::numeric / nullif(ae.completed_appointments, 0)::numeric)::numeric, 2)
  * Root cause: PostgreSQL round() with precision requires numeric type, but division with avg() (double precision) or integer types produces double precision

- Remove PII column tests from _mart_hygiene_retention.yml
  * Remove provider_name and hygienist_name tests (columns removed due to PII concerns)

- Update recommended_recall_interval accepted_values test
  * Add 24 months to accepted values list [3, 4, 5, 6, 8, 12, 18, 24]
  * 24 months is valid for full mouth X-rays and panoramic X-rays (2-year recall interval)
  * Fixes test failure affecting 1,287 records

- Add .cursorrules to enforce schema naming conventions
  * Prevent use of "public" schema in queries
  * Enforce use of raw, staging, int, and marts schemas

- Add investigation query for recall interval values
  * investigate_recall_interval_values.sql for debugging test failures

+++ dbt_dental_models/models/marts/mart_appointment_summary.sql
@@ -262,7 +262,7 @@ final as (
         ae.total_scheduled_production,
         ae.completed_production,
         round((ae.completed_production::numeric / nullif(ae.total_scheduled_production, 0) * 100)::numeric, 2) as production_efficiency,
-        round((ae.completed_production::numeric / nullif(ae.completed_appointments, 0))::numeric, 2) as avg_production_per_completed_appointment,
+        round((ae.completed_production::numeric / nullif(ae.completed_appointments, 0)::numeric)::numeric, 2) as avg_production_per_completed_appointment,
         
         -- Performance Categories (moved to enhanced CTE)
         ae.completion_performance,
