+++ dbt_dental_models/models/marts/mart_hygiene_retention.sql
@@ -152,7 +152,7 @@ hygiene_enhanced as (
         round(ha.hygiene_cancellations::numeric / nullif(ha.total_hygiene_appointments, 0) * 100, 2) as hygiene_cancellation_rate,
         round(ha.regular_interval_visits::numeric / nullif(ha.total_intervals, 0) * 100, 2) as regular_interval_percentage,
         round(ha.avg_hygiene_interval_days / 30.0, 1) as avg_hygiene_interval_months,
-        round(ha.avg_hygiene_production::numeric / nullif(ha.avg_hygiene_duration, 0) * 60, 2) as production_per_hour,
+        round((ha.avg_hygiene_production::numeric / nullif(ha.avg_hygiene_duration, 0) * 60)::numeric, 2) as production_per_hour,
         
         -- Days since last hygiene
         case when ha.last_hygiene_date is not null
